Phase T: GOP-first display backend — real-hardware portability milestone 2
Loader (uefi-boot/src/gop.rs): query the UEFI Graphics Output Protocol before
ExitBootServices (locate_handle_buffer + open_protocol_exclusive::<GraphicsOutput>),
prefer a 800x600x32 mode via a real firmware set_mode call (fall back to the
firmware's current mode — the kernel centers its 640x400 desktop in any
resolution), and write a fixed 40-byte handoff block at boot-info offset 5199
(GOP_OFFSET): present/pixel_format/width/height/stride_px/bpp/framebuffer_base/
framebuffer_size. Mirror struct on both sides.

Kernel: display seam is now a trait — gpu.rs GpuDevice over a GpuBackend enum
(GopGpu | BochsGpu); gpu_compositor::blit_cells<G: GpuDevice> writes BGRX or
RGBX byte order per the mode's bgr flag; desktop.rs install_gpu(GpuBackend);
main.rs selects GOP-first, Bochs-VBE probe as fallback. The legacy VGA text
console no longer kills the firmware's display mode: vga_init only blanks the
text buffer and vga_enter_text_mode (which disables Bochs VBE) runs only when
no pixel backend was installed — found live: the old vga_init unconditionally
disabled the VBE mode OVMF had set, leaving the GOP framebuffer invisible.

Honest limits documented: GOP framebuffer must sit below 4 GiB (loader
identity map covers 0..4 GiB; higher bases rejected with Bochs fallback —
common on discrete GPUs); still QEMU/OVMF-verified, not physical hardware;
no runtime mode re-probing.

Tests: kernel 579 -> 595 (10 boot_info GopHandoff + 5 gpu GopGpu + 1 compositor
byte-order), aegis 128, uefi-boot 22 -> 745 total featured; fmt/clippy clean
(also fixed 3 pre-existing unnecessary_cast warnings in uefi-boot elf.rs);
RUSTFLAGS="-Dwarnings" kernel release build clean.

Live evidence (qemu-gop-demo.ps1, serial-gop-demo.log, scr-gop-demo.ppm):
Aegis: GOP: set mode 800x600x32 -> handoff written at 5199 -> kernel reports
GOP framebuffer 800x600@32bpp (pitch 3200, BGRX byte order) at 0x80000000 ->
desktop renders through the GOP path; taskbar demo re-run on the GOP path:
4/4 taskbar@click serial lines, per-click screendumps pixel-identical to the
Phase-S Bochs-path evidence.

+++ uefi-boot/src/elf.rs
@@ -187,11 +187,11 @@ fn parse_relocations(data: &[u8]) -> Result<Vec<Relocation>, ElfError> {
     if shoff == 0 || shnum == 0 || shentsize < 64 {
         return Ok(Vec::new()); // No section headers: no relocations to apply
     }
-    if shstrndx as usize >= shnum as usize {
+    if shstrndx >= shnum {
         return Err(ElfError);
     }
 
-    let shstr_off = shoff + shstrndx as usize * shentsize;
+    let shstr_off = shoff + shstrndx * shentsize;
     let shstr_offset = u64::from_le_bytes(
         data[shstr_off + 24..shstr_off + 32]
             .try_into()
@@ -209,7 +209,7 @@ fn parse_relocations(data: &[u8]) -> Result<Vec<Relocation>, ElfError> {
 
     let mut relocations = Vec::new();
     for i in 0..shnum {
-        let sec_off = shoff + i as usize * shentsize;
+        let sec_off = shoff + i * shentsize;
         let name_idx = u32::from_le_bytes(
             data[sec_off..sec_off + 4]
                 .try_into()
